[async.core] changed: core interface

+++ process.py
@@ -6,7 +6,7 @@ import pickle
 import signal
 
 from .disposable import Disposable, CompositeDisposable
-from .async import Async, AsyncReturn, Future, SucceededFuture, Pipe, BrokenPipeError
+from .async import Async, AsyncReturn, Future, FutureSource, SucceededFuture, Pipe, BrokenPipeError
 
 __all__ = ('Process', 'ProcessCall', 'ProcessError', 'PIPE', 'DEVNULL', 'STDIN', 'STDOUT', 'STDERR',)
 #------------------------------------------------------------------------------#
@@ -26,26 +26,24 @@ class Process (object):
     """Create child process
     """
     def __init__ (self, command, stdin = None, stdout = None, stderr = None, preexec = None,
-            shell = None, environ = None, check = None, buffer_size = None, core = None):
-
+            shell = None, environ = None, check = None, buffer_size = None, kill = None, core = None):
         # vars
-        self.command     = ['/bin/sh', '-c', command] if shell else command
-        self.environ     = environ
-        self.check       = check is None or check
+        self.command = ['/bin/sh', '-c', command] if shell else command
+        self.environ = environ
+        self.kill    = kill
+        self.check   = check is None or check
 
         # dispose
         self.dispose = CompositeDisposable ()
 
         # state
         self.pid    = None
-        self.status = None
+        self.status = FutureSource ()
 
         #----------------------------------------------------------------------#
         # Pipes                                                                #
         #----------------------------------------------------------------------#
-        stdin_fd = self.to_fd (stdin, STDIN)
-        self.stdin_pipe = self.dispose.Add (
-            Pipe (None if stdin_fd is None else (stdin_fd, None), buffer_size, core))
+        status_pipe  = self.dispose.Add (Pipe (core = core))
 
         stdout_fd = self.to_fd (stdout, STDOUT)
         self.stdout_pipe = self.dispose.Add (
@@ -55,7 +53,9 @@ class Process (object):
         self.stderr_pipe = self.dispose.Add (
             Pipe (None if stderr_fd is None else (None, stderr_fd), buffer_size ,core))
 
-        status_pipe  = self.dispose.Add (Pipe (core = core))
+        stdin_fd = self.to_fd (stdin, STDIN)
+        self.stdin_pipe = self.dispose.Add (
+            Pipe (None if stdin_fd is None else (stdin_fd, None), buffer_size, core))
 
         #----------------------------------------------------------------------#
         # Fork                                                                 #
@@ -75,32 +75,25 @@ class Process (object):
                         stream.CloseOnExec (True)
 
             # status
-            self.status = self.status_main (status_pipe.Read)
+            @Async
+            def status_main ():
+                try:
+                    error_dump = yield error_stream.ReadUntilEof ()
+                    if error_dump:
+                        self.status.ErrorRaise (pickle.loads (error_dump))
+                except BrokenPipeError: pass
+                finally:
+                    self.Dispose ()
+            status_main ()
 
         else:
             try:
-                # status
-                status_pipe.Write.Blocking (True)
-                status_fd = status_pipe.Write.Detach ().Result ()
-                status_pipe.Dispose ()
+                # status descriptor (must not block)
+                status_fd = status_pipe.DetachWrite ().Result ()
 
-                # standard input
-                self.stdin_pipe.Read.Blocking (True)
-                if self.stdin_pipe.Read.Fd != 0:
-                    os.dup2 (self.stdin_pipe.Read.Fd, 0)
-                    self.stdin_pipe.Dispose ()
-
-                # standard output
-                self.stdout_pipe.Write.Blocking (True)
-                if self.stdout_pipe.Write.Fd != 1:
-                    os.dup2 (self.stdout_pipe.Write.Fd, 1)
-                    self.stdin_pipe.Dispose ()
-
-                # standard error
-                self.stderr_pipe.Write.Blocking (True)
-                if self.stderr_pipe.Write.Fd != 2:
-                    os.dup2 (self.stderr_pipe.Write.Fd, 2)
-                    self.stderr_pipe.Dispose ()
+                self.stdin_pipe.DetachRead (0)
+                self.stdout_pipe.DetachWrite (1)
+                self.stderr_pipe.DetachWrite (2)
 
                 # pre-exec hook
                 if preexec is not None:
@@ -141,7 +134,7 @@ class Process (object):
     def Status (self):
         """Future object for return code of the process
         """
-        return self.status
+        return self.status.Future
 
     @property
     def Pid (self):
@@ -164,9 +157,18 @@ class Process (object):
         finally:
             error_stream.Dispose ()
 
+
+            '''
+            while True:
+                pid, status = os.waitpid (self.pid, os.WNOHANG)
+                if pid != self.pid:
+                    yield error_stream.Core.WhenTimeDelay (1)
+            '''
+
             pid, status = os.waitpid (self.pid, os.WNOHANG)
             if pid != self.pid:
-                os.kill (self.pid, signal.SIGTERM)
+                if self.kill:
+                    os.kill (self.pid, signal.SIGTERM)
                 pid, status = os.waitpid (self.pid, 0)
 
             if error_dump:
@@ -213,7 +215,8 @@ class Process (object):
 # Call Process                                                                 #
 #------------------------------------------------------------------------------#
 def ProcessCall (command, input = None, stdin = None, stdout = None, stderr = None,
-        shell = None, environ = None, check = None, buffer_size = None, core = None, cancel = None):
+    preexec = None, shell = None, environ = None, check = None, buffer_size = None,
+    kill = None, core = None, cancel = None):
     """Asynchronously run command
 
     Asynchronously returns standard output, standard error and return code tuple.
@@ -230,15 +233,18 @@ def ProcessCall (command, input = None, stdin = None, stdout = None, stderr = No
     # process helper
     @Async
     def process ():
-        with Process (command, stdin, stdout, stderr, shell, environ, check, buffer_size, core) as proc:
+        with Process (command = command, stdin = stdin, stdout = stdout, stderr = stderr,
+            preexec = preexec, shell = shell, environ = environ, check = check,
+            buffer_size = buffer_size, kill = kill, core = core) as proc:
+
             # input
             if input:
                 proc.Stdin.Write (input)
                 proc.Stdin.Flush ().Continue (lambda *_: proc.Stdin.Dispose ())
 
             # output
-            out = proc.Stdout.ReadUntilEof () if proc.Stdout else SucceededFuture (None)
-            err = proc.Stderr.ReadUntilEof () if proc.Stderr else SucceededFuture (None)
+            out = proc.Stdout.ReadUntilEof (cancel) if proc.Stdout else SucceededFuture (None)
+            err = proc.Stderr.ReadUntilEof (cancel) if proc.Stderr else SucceededFuture (None)
             yield Future.WhenAll ((out, err))
 
             AsyncReturn ((out.Result (), err.Result (), (yield proc.Status)))
